Fix an int narrowing warning on 32-bit

+++ src/impl/transact_log_handler.cpp
@@ -55,7 +55,7 @@ struct MarkDirtyMixin  {
     bool set_link(size_t c, size_t r, size_t, size_t, _impl::Instruction i) { return mark_dirty(r, c, i); }
     bool set_null(size_t c, size_t r, _impl::Instruction i, size_t) { return mark_dirty(r, c, i); }
 
-    bool add_int(size_t col, size_t row, size_t) { return mark_dirty(row, col); }
+    bool add_int(size_t col, size_t row, int_fast64_t) { return mark_dirty(row, col); }
     bool nullify_link(size_t col, size_t row, size_t) { return mark_dirty(row, col); }
     bool insert_substring(size_t col, size_t row, size_t, StringData) { return mark_dirty(row, col); }
     bool erase_substring(size_t col, size_t row, size_t, size_t) { return mark_dirty(row, col); }
